Add inspect element to menu if isDev - fixes #1
This adds an "Inspect Element" item to the end of the menu if the app is in `development`, which opens the web inspector and focuses it on the element under the mouse when the menu is opened. Basically like how Chrome works.

The only reason I added the `electron` package is because I'm getting a `xo` error without it. I don't mind removing it, but I think I'd have to add a comment to ignore the `require` line in `index.js`. Could be wrong, feel free to let me know.

This fixes #1.

Closes #4